hacoMobile: adding various project dependencies and resources
