+++ HacoMobile/app/src/main/AndroidManifest.xml
@@ -4,22 +4,17 @@
 
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
+        android:icon="@mipmap/ic_haco_logo"
+        android:label="@string/str_app_name"
+        android:roundIcon="@mipmap/ic_haco_logo_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-
-        <activity android:name=".SplashActivity" android:theme="@style/SplashTheme">
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-
-        <activity android:name=".MainActivity">
-
         </activity>
         
     </application>
