--- HacoMobile/app/src/main/java/com/example/haco/DBHandler.kt
@@ -1,5 +1,0 @@
-package com.example.haco
-
-class DBHandler {
-    
-}
