--- HacoMobile/app/src/main/java/com/example/haco/MainActivity.kt
@@ -1,11 +1,0 @@
-package com.example.haco
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
