+++ HacoMobile/app/src/main/java/com/example/haco/activities/Login.kt
@@ -1,0 +1,13 @@
+package com.example.haco.activities
+
+import android.os.Bundle
+import android.os.PersistableBundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.haco.R
+
+class Login : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.login)
+    }
+}
