+++ HacoMobile/app/src/main/java/com/example/haco/activities/Profile.kt
@@ -1,0 +1,7 @@
+package com.example.haco.activities
+
+import androidx.appcompat.app.AppCompatActivity
+
+class Profile : AppCompatActivity() {
+
+}
