+++ HacoMobile/app/src/main/java/com/example/haco/activities/SplashActivity.kt
@@ -1,9 +1,8 @@
-package com.example.haco
+package com.example.haco.activities
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat.startActivity
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?){
