+++ HacoMobile/app/src/main/java/com/example/haco/helpers/DBHandler.kt
@@ -1,0 +1,5 @@
+package com.example.haco.helpers
+
+class DBHandler {
+    
+}
