+++ HacoMobile/app/src/main/res/layout/activity_main.xml
@@ -4,18 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Hello World!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.591" />
+    tools:context=".activities.MainActivity">
 
     <LinearLayout
         android:id="@+id/tb_haco"
@@ -30,25 +19,64 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <ImageView
-        android:id="@+id/img_hacoLogo"
-        android:layout_width="204dp"
-        android:layout_height="208dp"
-        android:layout_marginStart="100dp"
-        android:layout_marginTop="120dp"
-        android:contentDescription="@string/app_name"
+    <com.google.android.gms.maps.MapView
+        android:id="@+id/map_HomeScreen"
+        android:layout_width="326dp"
+        android:layout_height="490dp"
+        android:layout_marginStart="44dp"
+        android:layout_marginTop="136dp"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:srcCompat="@tools:sample/backgrounds/scenic" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/txt_hacoWelcome"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="176dp"
-        android:layout_marginTop="42dp"
-        android:text="@string/app_name"
+        android:id="@+id/txtV_Haco"
+        android:layout_width="213dp"
+        android:layout_height="39dp"
+        android:layout_marginStart="99dp"
+        android:layout_marginTop="28dp"
+        android:gravity="center"
+        android:text="@string/str_app_name"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/img_hacoLogo" />
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:id="@+id/nav_bottomHome"
+        android:layout_width="408dp"
+        android:layout_height="61dp"
+        android:layout_marginStart="1dp"
+        android:layout_marginBottom="1dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent" >
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="horizontal">
+
+            <Button
+                android:id="@+id/btn_nav_login"
+                style="@style/Widget.AppCompat.Button"
+                android:layout_width="110dp"
+                android:layout_height="match_parent"
+                android:text="@string/str_login" />
+
+            <Button
+                android:id="@+id/btn_nav_report"
+                style="@style/Widget.AppCompat.Button"
+                android:layout_width="110dp"
+                android:layout_height="match_parent"
+                android:layout_gravity="bottom"
+                android:text="@string/str_report" />
+
+            <Button
+                android:id="@+id/btn_nav_profile"
+                style="@style/Widget.AppCompat.Button"
+                android:layout_width="0dp"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:text="@string/str_profile" />
+
+        </LinearLayout>
+    </com.google.android.material.bottomnavigation.BottomNavigationView>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
