+++ HacoMobile/app/src/main/res/values/strings.xml
@@ -1,3 +1,7 @@
 <resources>
-    <string name="app_name">Haco</string>
+    <string name="str_app_name">Haco</string>
+    <string name="str_login">Login</string>
+    <string name="str_register">Register</string>
+    <string name="str_profile">Profile</string>
+    <string name="str_report">Report</string>
 </resources>
